Refactores column definitions. Now they are included in each component page, to make it more maintainable
