+++ src/app/ui/categories-table.tsx
@@ -1,13 +1,84 @@
 'use client';
 
-import { useState } from 'react';
-import { columns } from "@/app/ui/tables/categories-columns"
+import { ColumnDef } from "@tanstack/react-table"
+import { ArrowUpDown, MoreHorizontal } from "lucide-react"
+import { Button } from "@/components/ui/button"
 import { DataTable } from "@/app/ui/data-table"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+
+type Category = {
+  category_id: string;
+  category_name: string;
+  category_description: string;
+}
+
+async function deleteCategory(categoryId: string) {
+  // TODO: Implement delete functionality
+  console.log('Deleting category:', categoryId);
+}
+
+const columns: ColumnDef<Category>[] = [
+  {
+    accessorKey: "category_name",
+    header: ({ column }) => {
+      return (
+        <Button
+          variant="ghost"
+          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
+        >
+          Name
+          <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      )
+    },
+  },
+  {
+    accessorKey: "all_columns",
+    header: "",
+    cell: ({ row }) => {
+      // This table is super simple so the "all_columns" is just the category_name
+      return <input type="hidden" value={row.original.category_name} />
+    },
+  },
+  {
+    id: "actions",
+    cell: ({ row }) => {
+      const category = row.original
+      return (
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button variant="ghost" className="h-8 w-8 p-0">
+              <span className="sr-only">Open menu</span>
+              <MoreHorizontal className="h-4 w-4" />
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent align="end">
+            <DropdownMenuLabel>Actions</DropdownMenuLabel>
+            <DropdownMenuItem
+              onClick={() => deleteCategory(category.category_id)}
+            >
+              Delete category
+            </DropdownMenuItem>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem>Edit category</DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      )
+    },
+  }
+];
 
 export default function CategoriesTable({ 
   categories 
 }: { 
-  categories: any[]
+  categories: Category[]
 }) {
   return (
     <div className="container mx-auto py-10">
