+++ src/app/ui/persons-table.tsx
@@ -1,12 +1,103 @@
 'use client';
 
-import { columns } from "@/app/ui/tables/persons-columns"
+import { ColumnDef } from "@tanstack/react-table"
+import { ArrowUpDown, MoreHorizontal } from "lucide-react"
+import { Button } from "@/components/ui/button"
 import { DataTable } from "@/app/ui/data-table"
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuLabel,
+  DropdownMenuSeparator,
+  DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu"
+
+type Person = {
+  person_id: string;
+  person_name: string;
+  person_surname: string;
+  person_email: string;
+  all_columns: string;
+}
+
+async function deletePerson(personId: string) {
+  // TODO: Implement delete functionality
+  console.log('Deleting person:', personId);
+}
+
+const columns: ColumnDef<Person>[] = [
+  {
+    accessorKey: "person_name",
+    header: ({ column }) => {
+      return (
+        <Button
+          variant="ghost"
+          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
+        >
+          Name
+          <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      )
+    },
+  },
+  {
+    accessorKey: "person_surname",
+    header: ({ column }) => {
+      return (
+        <Button
+          variant="ghost"
+          onClick={() => column.toggleSorting(column.getIsSorted() === "asc")}
+        >
+          Surname
+          <ArrowUpDown className="ml-2 h-4 w-4" />
+        </Button>
+      )
+    },
+  },
+  {
+    accessorKey: "person_email",
+    header: "Email",
+  },
+  {
+    accessorKey: "all_columns",
+    header: "",
+    cell: ({ row }) => {
+      return <input type="hidden" value={row.getValue("all_columns")} />
+    },
+  },
+  {
+    id: "actions",
+    cell: ({ row }) => {
+      const person = row.original
+      return (
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button variant="ghost" className="h-8 w-8 p-0">
+              <span className="sr-only">Open menu</span>
+              <MoreHorizontal className="h-4 w-4" />
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent align="end">
+            <DropdownMenuLabel>Actions</DropdownMenuLabel>
+            <DropdownMenuItem
+              onClick={() => deletePerson(person.person_id)}
+            >
+              Delete person
+            </DropdownMenuItem>
+            <DropdownMenuSeparator />
+            <DropdownMenuItem>Edit person</DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
+      )
+    },
+  }
+];
 
 export default function PersonsTable({ 
   persons 
 }: { 
-  persons: any[]
+  persons: Person[]
 }) {
   return (
     <div className="container mx-auto py-10">
